refactor: add `core` reexport
