+++ lib/index.ts
@@ -6,8 +6,5 @@
 export type { ConflictResolutionResult, ConflictPickerOptions } from './ui/ui.ts'
 export { openConflictPicker, uploadConflictHandler, UploadPicker } from './ui/ui.ts'
 
-export type { IDirectory, Directory } from './core/utils/fileTree.ts'
-export { getConflicts, hasConflict } from './core/utils/conflicts.ts'
-export { Upload, Status as UploadStatus } from './core/upload.ts'
-export * from './core/uploader'
-export { getUploader, upload } from './core/getUploader.ts'
+export type { Directory, IDirectory, Eta, EtaEventsMap } from './core/core.ts'
+export { EtaStatus, getConflicts, getUploader, hasConflict, upload, Upload, UploadStatus, Uploader, UploaderStatus } from './core/core.ts'
